Stochastically sampled occupancy convolution
A sampled convolution that utilises simple random sampling is implemented.

- Move *kernelNormalisationFactor* from `Full::Base` to `Base`, and rename to *area*.
- Promote `KernelSizeType` to 32-bit to avoid overflow when calculating area of a large kernel.
- Make `std::ranges::transform` functor unconditionally `noexcept`, because STL views are always nothrow viewable.
- Improve cache temporal locality by clearing the kernel memory right before it is being written.
- Sampled.Systematic: Default initialise its members.
- Core.XXHash:
  - Promote `RandomEngine::CounterType` to 32-bit to avoid overflow when generating a long sequence.
  - Add more static assertion on the choice of random bit generator for secret.
- RegionfieldGenerator.VoronoiGenerator: Assignment of region identifier to the centroid is now conditionally dependent on the centroid coordinate.

+++ Library/DisRegRep/Core/XXHash.hpp
@@ -137,7 +137,7 @@ public:
 	using ResultType = HashType;
 	using ResultLimit = std::numeric_limits<ResultType>;
 
-	using CounterType = std::uint_fast16_t;
+	using CounterType = std::uint_fast32_t;
 	using CounterReference = std::add_lvalue_reference_t<CounterType>;
 	using ObjectType = std::tuple<CounterType, Obj...>;
